@@ -13,11 +13,11 @@ group = 'com.github.MatrixFrog'
 version = '1.0.4'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 31
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 30
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
     }
@@ -35,7 +35,7 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:30.0.0'
+    compile 'com.android.support:appcompat-v7:31.0.0'
 }
 
 afterEvaluate {
